refactor: extract useItemListState + useBatchActions from item-list (TD-015)
Extract 3 modules from 578-line item-list.tsx:
- src/lib/batch-actions.tsx: BatchActionConfig type + getBatchActions()
- src/hooks/use-item-list-state.ts: all list state (queries, pagination,
  sorting, selection, category grouping)
- src/hooks/use-batch-actions.ts: batch mutation + handleBatchAction

item-list.tsx reduced from 578 to 257 lines (rendering only).
Merged duplicate handleSelect/handleNavigate into navigateToItem.
Extracted renderItemCard helper to eliminate 3x repeated ItemCard JSX.
